Split models into SMILES-specific and general
